Task 02_01 (Level C) Burov

+++ src/by/it/burov/jd02_01/utils/Constants.java
@@ -1,5 +1,5 @@
 package by.it.burov.jd02_01.utils;
 
 public class Constants {
-    public static final int K_SPEED = 100;
+    public static final double K_SPEED = 100;
 }
